@@ -1,90 +1,89 @@
 {
-    "name": "laytrip-website",
-    "version": "0.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build:dev": "ng build",
-        "build:suresh": "ng serve --configuration=suresh",
-        "build:prod": "ng build --prod",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e",
-        "compile:server": "webpack --config webpack.server.config.js --progress --colors",
-        "serve:ssr": "node dist/server",
-        "build:ssr": "npm run build:client-and-server-bundles && npm run compile:server",
-        "build:client-and-server-bundles": "ng build --prod --outputHashing=all && ng run laytrip-website:server:production"
-    },
-    "private": true,
-    "dependencies": {
-        "@agm/core": "1.0.0-beta.7",
-        "@agm/js-marker-clusterer": "^1.1.0",
-        "@angular/animations": "~7.2.0",
-        "@angular/cdk": "~7.3.7",
-        "@angular/common": "~7.2.0",
-        "@angular/compiler": "~7.2.0",
-        "@angular/core": "~7.2.0",
-        "@angular/forms": "~7.2.0",
-        "@angular/http": "~7.2.0",
-        "@angular/platform-browser": "~7.2.0",
-        "@angular/platform-browser-dynamic": "~7.2.0",
-        "@angular/platform-server": "^7.2.16",
-        "@angular/router": "~7.2.0",
-        "@ng-bootstrap/ng-bootstrap": "^4.0.0",
-        "@ng-select/ng-select": "^2.20.5",
-        "@nguniversal/express-engine": "^7.1.1",
-        "@nguniversal/module-map-ngfactory-loader": "0.0.0",
-        "@ngx-translate/core": "^13.0.0",
-        "@ngx-translate/http-loader": "^6.0.0",
-        "@types/google-maps": "^3.2.2",
-        "@types/googlemaps": "^3.40.3",
-        "angular2-text-mask": "^9.0.0",
-        "angularx-social-login": "^3.2.2",
-        "bootstrap": "^4.5.0",
-        "collect.js": "^4.28.6",
-        "core-js": "^2.5.4",
-        "express": "^4.15.2",
-        "hammerjs": "^2.0.8",
-        "jquery": "^3.5.1",
-        "js-marker-clusterer": "^1.0.0",
-        "jwt-decode": "^2.2.0",
-        "localstorage-polyfill": "^1.0.1",
-        "moment": "^2.27.0",
-        "ng-lottie": "^0.3.2",
-        "ng5-slider": "^1.2.4",
-        "ngx-cookie": "^5.0.0",
-        "ngx-countdown": "^3.1.0",
-        "ngx-gallery": "^5.10.0",
-        "ngx-toastr": "^10.1.0",
-        "node-sass": "^4.0.0",
-        "popper.js": "^1.16.1",
-        "primeng": "7.1.0",
-        "rxjs": "^6.3.3",
-        "rxjs-compat": "^6.6.2",
-        "tslib": "^1.9.0",
-        "zone.js": "~0.8.26"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "~0.13.0",
-        "@angular/cli": "~7.3.10",
-        "@angular/compiler-cli": "~7.2.0",
-        "@angular/language-service": "~7.2.0",
-        "@types/node": "~8.9.4",
-        "@types/jasmine": "~2.8.8",
-        "@types/jasminewd2": "~2.0.3",
-        "codelyzer": "~4.5.0",
-        "jasmine-core": "~2.99.1",
-        "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~4.0.0",
-        "karma-chrome-launcher": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "~2.0.1",
-        "karma-jasmine": "~1.1.2",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "protractor": "~5.4.0",
-        "ts-loader": "^5.2.0",
-        "ts-node": "~7.0.0",
-        "tslint": "~5.11.0",
-        "typescript": "~3.2.2",
-        "webpack-cli": "^3.1.0"
-    }
+  "name": "laytrip-website",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build:dev": "ng build",
+    "build:suresh": "ng serve --configuration=suresh",
+    "build:prod": "ng build --prod",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "compile:server": "webpack --config webpack.server.config.js --progress --colors",
+    "serve:ssr": "node dist/server",
+    "build:ssr": "npm run build:client-and-server-bundles && npm run compile:server",
+    "build:client-and-server-bundles": "ng build --prod --outputHashing=all && ng run laytrip-website:server:production"
+  },
+  "private": true,
+  "dependencies": {
+    "@agm/core": "1.0.0-beta.7",
+    "@agm/js-marker-clusterer": "^1.1.0",
+    "@angular/animations": "~8.2.14",
+    "@angular/cdk": "~7.3.7",
+    "@angular/common": "~8.2.14",
+    "@angular/compiler": "~8.2.14",
+    "@angular/core": "~8.2.14",
+    "@angular/forms": "~8.2.14",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/platform-server": "^8.2.14",
+    "@angular/router": "~8.2.14",
+    "@ng-bootstrap/ng-bootstrap": "^4.0.0",
+    "@ng-select/ng-select": "^2.20.5",
+    "@nguniversal/express-engine": "^7.1.1",
+    "@nguniversal/module-map-ngfactory-loader": "0.0.0",
+    "@ngx-translate/core": "^13.0.0",
+    "@ngx-translate/http-loader": "^6.0.0",
+    "@types/google-maps": "^3.2.2",
+    "@types/googlemaps": "^3.40.3",
+    "angular2-text-mask": "^9.0.0",
+    "angularx-social-login": "^3.2.2",
+    "bootstrap": "^4.5.0",
+    "collect.js": "^4.28.6",
+    "core-js": "^2.5.4",
+    "express": "^4.15.2",
+    "hammerjs": "^2.0.8",
+    "jquery": "^3.5.1",
+    "js-marker-clusterer": "^1.0.0",
+    "jwt-decode": "^2.2.0",
+    "localstorage-polyfill": "^1.0.1",
+    "moment": "^2.27.0",
+    "ng-lottie": "^0.3.2",
+    "ng5-slider": "^1.2.4",
+    "ngx-cookie": "^5.0.0",
+    "ngx-countdown": "^3.1.0",
+    "ngx-gallery": "^5.10.0",
+    "ngx-toastr": "^10.1.0",
+    "node-sass": "^4.0.0",
+    "popper.js": "^1.16.1",
+    "primeng": "10.0.0",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.6.2",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular/cli": "~8.3.29",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/language-service": "~8.2.14",
+    "@types/node": "~8.9.4",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "~5.4.0",
+    "ts-loader": "^5.2.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.5.3",
+    "webpack-cli": "^3.1.0"
+  }
 }
